[USA] Tracelog Capture - First full playthrough of the game
98.208% Marked (1,029,787/1,048,576 bytes)

First full playthough of the game. A majority of the game's code is marked. Of the remaining bytes, I believe half belong to data; in particular, $81CD9C through $81F400 might all be a huge chunk of data. Will need to verify this for sure. I know of a few things that remain to do (use mouse/multitap), but from there, it might be time to manually start marking things.